+++ R/chronon_convert.R
@@ -28,23 +28,21 @@
 chronon_convert <- S7::new_generic("chronon_convert", "x")
 
 chronon_convert_impl <- function(x, from, to, discrete, tz = NULL) {
-  # Convert between identical time units
-  if (identical(S7::S7_class(from), S7::S7_class(to))) {
-    x <- vec_data(x) * from@n / to@n
-    if (discrete) x <- as.integer(floor(x))
-    return(x)
-  }
-
   # Inherit granule properties from `from` if not given in `to` granule
   to <- granule_inherit_props(to, from)
 
   # Handle timezones
   # TODO - Generalise for converting to any different granule attribute
   # (e.g. locA -> locB)
-  if (is.null(tz)) {
-    tz <- tz_name(to)
-    # UTC is a suitable proxy for naive time zones
-    if (is.na(tz)) tz <- "UTC"
+
+  is_tz_conv <- S7::S7_inherits(from, mt_tz_unit) && S7::S7_inherits(to, mt_tz_unit)
+  if (is_tz_conv) {
+    if (is.na(tz_name(from)) && !is.na(tz_name(to))) {
+      cli::cli_abort(
+        "Cannot convert from timezone-naive chronon to timezone-aware chronon.",
+        call = NULL
+      )
+    }
   }
   
   # Find path along convertable time units
@@ -59,38 +57,56 @@ chronon_convert_impl <- function(x, from, to, discrete, tz = NULL) {
     to
   )
 
-  # Find timezone offsets for chronon boundaries
-  if (tz != "UTC") {
-    `1s` <- cal_time_civil$second(1L)
-    xs <- chronon_convert_impl(x, from, `1s`, discrete = FALSE, tz = "UTC")
-    xso <- get_tz_offset(as.double(xs), tz)
-    x <- chronon_convert_impl(xs + xso, `1s`, from, discrete = FALSE, tz = "UTC")
+  # Convert to target tz naive time for standard chronon boundaries
+  # Skipped if:  
+  # * target tz is UTC (since that matches the internal representation)
+  is_tz_utc <- identical(tz_name(to), "UTC") || identical(tz, "UTC")
+  # * target is not a timezone-aware mt_tz_unit
+  is_tz_aware <- is_tz_conv && !is.na(tz_name(to))
+
+  if (!is_tz_utc && is_tz_aware) {
+    tzo <- tz_offset_impl(x, chronon = from, tz = tz_name(to))
+    x <- x + tzo
+  } else if (is_tz_conv && is.na(tz_name(to)) && !is.na(tz_name(from))) {
+    # Converting from tz-aware to tz-naive - remove offset
+    tzo <- tz_offset_impl(x, chronon = from, tz = tz_name(from))
+    x <- x + tzo
   }
 
-  path[[1]] <- from
-  prop_from <- S7::prop_names(from)
-  path[[length(path)]] <- to
-  prop_to <- S7::prop_names(to)
-  # Initialise intermediate classes with 1L and adjacent properties
-  path[c(-1, -length(path))] <- lapply(path[c(-1, -length(path))], function(tu){
-    # Ideally inherit from `to`, but if incomplete inherit from `from`
-    granule_inherit_props(granule_inherit_props(tu(1L), to), from)
-  })
-
-  # Convert chronons along the path
-  not_na <- !is.na(x)
-  for (i in seq(2, length.out = length(path)-1)) {
-    res <- chronon_divmod_dispatch(path[[i-1L]], path[[i]], x[not_na])
-    x[not_na] <- res$div
-    nz_mod <- res$mod != 0
-    part <- chronon_cardinality(path[[i-1L]], path[[i]], floor(res$div[nz_mod]))
-    x[not_na][nz_mod] <- x[not_na][nz_mod] + res$mod[nz_mod]/part
+  # Convert between time points
+  if (length(path) == 1L) {
+    # Simple direct conversion by cardinality
+    x <- vec_data(x) / chronon_cardinality(from, to, at = x)
+  } else {
+    # Convert along the path of time granules
+    path[[1]] <- from
+    prop_from <- S7::prop_names(from)
+    path[[length(path)]] <- to
+    prop_to <- S7::prop_names(to)
+    # Initialise intermediate classes with 1L and adjacent properties
+    path[c(-1, -length(path))] <- lapply(path[c(-1, -length(path))], function(tu){
+      # Ideally inherit from `to`, but if incomplete inherit from `from`
+      granule_inherit_props(granule_inherit_props(tu(1L), to), from)
+    })
+  
+    # Convert chronons along the path
+    not_na <- !is.na(x)
+    for (i in seq(2, length.out = length(path)-1)) {
+      res <- chronon_divmod_dispatch(path[[i-1L]], path[[i]], x[not_na])
+      x[not_na] <- res$div
+      nz_mod <- res$mod != 0
+      part <- chronon_cardinality(path[[i-1L]], path[[i]], floor(res$div[nz_mod]))
+      x[not_na][nz_mod] <- x[not_na][nz_mod] + res$mod[nz_mod]/part
+    }
   }
 
-  # Convert back to UTC time internally
-  if (tz != "UTC") {
+  # Convert back to absolute time for time zoned data
+  # Skipped if target tz is:
+  # * NA (since that indicates a timezone-naive chronon)
+  # * UTC (since that matches the internal representation)
+  if (!is_tz_utc && is_tz_aware) {
     f <- if (discrete) as.integer else identity
-    x <- x - f(chronon_convert_impl(xso, `1s`, to, discrete = FALSE, tz = "UTC"))
+    x <- x - f(chronon_convert_impl(tzo, from, to, discrete = FALSE, tz = "UTC"))
   }
 
   if (discrete) x <- as.integer(floor(x))