+++ NEWS.md
@@ -7,6 +7,7 @@
   (note: use `time_round()`, `time_floor()`, and `time_ceiling()` for rounding
   to a specified time granule).
 * Added casting from time durations to character vectors.
+* Added explicit conversion to naive time by using `tz = NA`.
 
 ## Improvements